+++ garbage_date_check/sunahara.rb
@@ -1,27 +1,27 @@
-# 西区浅間一丁目のゴミ収集日のデータ
-class GarbageDateSen1
+# 西区砂原町のゴミ収集日のデータ
+class GarbageDateSun
   require 'date'
-  require_relative 'date_patch'
+  require './date_patch'
   
   def self.target_day
     @target_day ||= Date.tomorrow
   end
 
   def self.burnable_garbage? #可燃ゴミ
-    target_day.monday? || target_day.thursday?
+    target_day.tuesday? || target_day.friday?
   end
 
   def self.not_burnable_garbage? #不燃ゴミ
     now = target_day
-    now.wednesday? && now.mweek == 4
+    now.wednesday? && now.mweek == 3
   end
 
   def self.plastic_garbage? #プラスチックゴミ
-    target_day.friday?
+    target_day.wednesday?
   end
 
   def self.bottle_can_garbage? #缶・ビン・ペットボトル
-    target_day.wednesday?
+    target_day.thursday?
   end
 
   def self.check
